// src/users/item/authentication/operations/item/longRunningOperationItemRequestBuilderDeleteRequestConfiguration.ts
import type {RequestOption} from '@microsoft/kiota-abstractions';

export interface LongRunningOperationItemRequestBuilderDeleteRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}