// src/users/item/authentication/operations/item/longRunningOperationItemRequestBuilderPatchRequestConfiguration.ts
import type {RequestOption} from '@microsoft/kiota-abstractions';

export interface LongRunningOperationItemRequestBuilderPatchRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}